+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 - Added `eraseStackTrace()` method in `EncodableResult` & `WrappedException` which erases the whole stack trace in exception information
 ### Fixed Bugs
 `reflection` module:
-- .decodeObject() not working properly with Java `Void`
-- .boxed()/.unboxed() not working properly with Java `Void`
+- `.encodeObject()`/`.decodeObject()` not working properly with `null`
+- `.decodeObject()` not working properly with Java `Void`
+- `.boxed()`/`.unboxed()` not working properly with Java `Void`